% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adv_example.R
\docType{data}
\name{adv_example}
\alias{adv_example}
\title{A toy dataset for the example to illustrate the possibility of multiple fire sale equilibria.}
\format{
A 1 x 4 tibble with the following variables
\describe{
  \item{Country}{Usually an ISO code. Here the country is called Thomas}
  \item{Year}{A number. We use 2020 because this is the year in which we found the example}
  \item{Volume}{A real number calibrated such that we match the impact function of the example}
  \item{Unit}{The fictious unit in which we measure the toy volume}
  \item{Currency}{The ficticious currency in which we measure the the toy volume}

}
}
\source{
\url{Breuer, T. and Summer M., and Urosevic B., "Systemic Loss Evaluation", A3}
}
\usage{
adv_example
}
\description{
This is a toy dataset for the example in Breuer, Summer, Urosevic, Appendix A3 given in a format
that it can be processed by the standard functions of the package which are also used for computing
the actual eba data
}
\keyword{datasets}
