% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/average_daily_volume_sovereign.R
\docType{data}
\name{average_daily_volume_sovereign}
\alias{average_daily_volume_sovereign}
\title{Average daily volume in Million Euro of sovereign bonds of DE, ES, GB, FR, IT, JP, US and the rest
of the world.}
\format{
A data frame with 88 rows and 5 variables:
\describe{
  \item{Country}{Country, ISO-code of the country.The rest of the world all excluding DE, ES, FR, IT, JP, UK, US is calles Total}
  \item{Year}{Year, year of the observation}
  \item{Volume}{Volume, Average daily turnover volume in Million Euro}
  \item{Unit}{Unit, unit of Volume}
  \item{Currency}{Currency, currency of volume}
}
}
\source{
The volume data are hand-collected from various sources on the internet
        DE \url{https://www.deutsche-finanzagentur.de/en/institutional-investors/secondary-market/}
        ES \url{https://www.tesoro.es/sites/default/files/estadisticas/15I.xlsx}
        FR \url{https://www.afme.eu/reports/data/details//Government-Bond-Data-Report-Q2-2019}
        IT \url{https://infostat.bancaditalia.it/}
        JP \url{https://asianbondsonline.adb.org/data-portal/}
        UK \url{https://www.dmo.gov.uk/data/gilt-market/turnover-data/}
        US \url{https://www.sifma.org/resources/research/us-treasury-trading-volume/}
        Total \url{https://www.bis.org/statistics/secstats.htm} and estimation of
        Cont and Schaanning 2017, equation (28)
}
\usage{
average_daily_volume_sovereign
}
\description{
A dataset containing the average daily volume of sovereign bonds for Germany, Spain, Great Britain,
France, Italy, Japan, United States and the rest of the world starting in 2009 up to the end of 2019.
Rest of the world: We have no data on the ADV in all the rest of exposures, because there are no
or insufficient direct observations. Cont and Schaanning (2016) observe a high correlation between
nominal debt outsanding and adv. They use this correlation by estimating unobserved ADV using regression
techniques. We apply an even simpler estimate. We compute the share of nominal debt outstanding of our
observed countries DE, ES, FR, IT, JP, GB, US in the total nominal government debt outstanding. We use this
proportionality factor to impute an ADV on the residual position. The dataset is compiled by running the
\code{make_price_volume_data} script in the data-raw folder. The exact compilation can be looked up in
this script.
}
\keyword{datasets}
