% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_sovereign_bond_price_volatility.R
\name{compute_sovereign_bond_price_volatility}
\alias{compute_sovereign_bond_price_volatility}
\title{compute_sovereign_bond_price_volatility}
\usage{
compute_sovereign_bond_price_volatility(data, from, to)
}
\arguments{
\item{data}{a dataframe/tibble containing the daily sovereign bond index values for all countries.}

\item{from}{start year of the data}

\item{to}{year up to which we consider the data}
}
\value{
sovereign bond log returns by country.
}
\description{
compute_sovereign_bond_price_volatility
}
\examples{
compute_sovereign_bond_price_volatility(sovereign_bond_indices, 2010, 2019)
}
