% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eba_exposures_2020.R
\docType{data}
\name{eba_exposures_2020}
\alias{eba_exposures_2020}
\title{Exposure data of 123 European banks which took part in the EBA 2020 transparency exercise.}
\format{
A data frame with 5628 rows and 11 variables:
\describe{
  \item{LEI_code}{LEI_code, legal entity identifyer of a bank.}
  \item{Country_code}{Country_code, ISO-code of the country where the bank is domiciled}
  \item{Bank_name}{Bank_name, Name of the bank}
  \item{Period}{Period, Reporting period of the data given as 201912 (31. December 2019)}
  \item{Country}{Country, Country to which the bank is exposed. This is either a ISO code like AT, DE etc. or it is
  called Total if it is the overall exposure aggregated across all countries, or there are special abbreviations
  which are further explained in the EBA metadata in the raw-data folder. They are not relevant for our analysis and
  not described further here.}
  \item{Exposure}{Asset class according to the IRB classification scheme}
  \item{Loan_Amount}{Loan_Amount, the exposure value in Million Euro held in the form of loans}
  \item{Bond_Amount}{Bond_Amount, the exposure value in Million Euro held as bonds}
  \item{Total_Amount}{Total_Amount, the exposure value as the sum of loan and bond amount in Million Euro}
  \item{Unit}{Unit, the units of Loan_Amount, Bond_Amount, Total_Amount}
  \item{Currency}{Currency, the currency of Loan_Amount, Bond_Amount, Total_Amount}
}
}
\source{
\url{https://eba.europa.eu/risk-analysis-and-data/eu-wide-transparency-exercise}
}
\usage{
eba_exposures_2020
}
\description{
A dataset containing the exposure data of 123 European banks who took part in the EBA transparency exercise 2020.
The actual number is 128 but we dropped five entities from the dataset because there were data problems.
The data are compiled from a raw data set. The script of compiling the data from the raw data can be
inspected by editing the
file make_balance_sheets_2020.R in the raw-data folder of the syslosseval package source code.
}
\keyword{datasets}
