% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fixed_point_computation_function.R
\name{fixed_point_computation_function}
\alias{fixed_point_computation_function}
\title{fixed_point_computation_function}
\usage{
fixed_point_computation_function(
  mat,
  lb,
  data_idx,
  data_adv,
  base_year,
  constant,
  accuracy = 10^(-9)
)
}
\arguments{
\item{mat}{A list of the initial state variable (output of
\code{make_state_variables()})}

\item{lb}{The critical leverage threshold called lambda_bar in the
paper.}

\item{data_idx}{The data-frame with the sovereign bond indices.}

\item{data_adv}{The data-frame with the average daily volume figures.}

\item{base_year}{The base year for the simulation}

\item{constant}{The value of the constant kappa in the impact fuction
(equation (9)).}

\item{accuracy}{The accuracy of the fixed point approximation. Set by
default to 10^9}
}
\value{
A tibble with variables delta_lower (lower fixed point), iterations_lower (iterations to
converge to lower fixed point), delta_upper (upper fixed point), iterations_uppper( iterations to
converge to the upper fixed point), delta_max (maximum impact), unique (logical variable which is
true if fixed point is unique and false if it is not unique)
}
\description{
This function computes the least and the greatest fire sale equilibrium
according to theorem 2 in the paper to a given level of accuracy, which is
set by default to 10^-9. In many cases the least and the greatest fire-sale
equilibrium will coincide for a given set of data, but this need not
generally be the case.
}
\examples{
stress_data <- make_stress_data(eba_exposures_2016, eba_impairments_2016, 1, 2015)
state_variables <- make_state_variables(stress_data)
fixed_point_computation_function(
  mat = state_variables, lb = 33,
  data_idx = sovereign_bond_indices,
  data_adv = average_daily_volume_sovereign,
  base_year = 2015,
  constant = 1.5
)
}
