% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fixed_point_computation_function_example.R
\name{fixed_point_computation_function_example}
\alias{fixed_point_computation_function_example}
\title{fixed_point_computation_function_example}
\usage{
fixed_point_computation_function_example(mat, lb, accuracy = 10^(-9))
}
\arguments{
\item{mat}{A list of the initial state variable}

\item{lb}{The critical leverage threshold called bar{lambda} in the paper.}

\item{accuracy}{The accuracy of the fixed point approximation. Set by default to 10^9}
}
\value{
A list with components equ_delta (equilibrium discount factor) and iter (number of iterations)
}
\description{
This function computes the lower and upper fixed point for the multiple equilibrium example in
appendix A3.
}
\examples{
fixed_point_computation_function_example(mat = example_multiple_equilibria, lb = 44)
}
