% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_price_impact_data.R
\name{make_price_impact_data}
\alias{make_price_impact_data}
\title{make_price_impact_data}
\usage{
make_price_impact_data(data_idx, data_adv, base_year)
}
\arguments{
\item{data_idx}{sovereign bond indices}

\item{data_adv}{average daily volume data}

\item{base_year}{the year for which the analysis is done}
}
\value{
A dataframe with variables: Country ADV_sov_bonds, sigma_sov_bonds.
}
\description{
make_price_impact_data
}
\examples{
make_price_impact_data(sovereign_bond_indices, average_daily_volume_sovereign, 2017)
}
