% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_S1.R
\name{make_S1}
\alias{make_S1}
\title{make_S1}
\usage{
make_S1(data)
}
\arguments{
\item{data}{a dataframe which is the output of \code{make_stress_data()}}
}
\value{
a B x I (number of banks x number of marketable security classes) with the sovereign
        bond exposures for DE, ES, IT, FR, JP, GB, US and Rest of the world
}
\description{
This function takes the dataframe which assembles the stresstest data from exposures and impairments, the
output of \code{make_stress_data()}, and constructs a matrix giving the sovereign bond exposures in
DE, ES, GB. FR, IT, JP, US and the rest of the world.
}
\examples{
stress_data <- make_stress_data(eba_exposures_2016, eba_impairments_2016, 1, 2015)
make_S1(stress_data)
}
