% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_sovereign_bond_exposures.R
\name{make_sovereign_bond_exposures}
\alias{make_sovereign_bond_exposures}
\title{make_sovereign_bond_exposures}
\usage{
make_sovereign_bond_exposures(data)
}
\arguments{
\item{data}{a dataframe with the compiled EBA exposure data}
}
\value{
a dataframe with the bond exposures for DE, ES, IT, FR, JP, GB, US and Rest of the world
}
\description{
make_sovereign_bond_exposures
}
\examples{
make_sovereign_bond_exposures(eba_exposures_2016)
}
