% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_state_variables.R
\name{make_state_variables}
\alias{make_state_variables}
\title{make_state_variables}
\usage{
make_state_variables(stress_data, meth = "asymmetric residuals")
}
\arguments{
\item{stress_data}{a dataframe which is the output of make_stress_data}

\item{meth}{a method, which can take the values "asymmetric residuals", "symmetric residuals".
It influences how the construction of state variables deals with the gaps between
EBA exposures and reported total assets.
We have to deal with the fact that the total value of EBA credit exposures is
for some banks larger than reported total assets and for some banks smaller.
This has to be corrected for computations of leverage. We suggest two
methods for correction. With the method "asymmetric residuals" we would compute
a residual value for the assets of each banks such that the total value of
EBA exposures for each bank plus the residual will add up to the value of
total assets reported in the bank balancesheet. In  the case where the value
of total assets is smaller than the value of EBA exposures the residual is
set to 0 so that the total value of EBA exposures is identified with the value of
total assets. In this case fro some banks leverage is perhaps overstated. With
the method "symmetric residuals" we compute the residual as the
difference between total assets as reported in the balance sheet and the total
value of eba exposures fro each bank. When we add this residual to the sum of
eba exposures we will get in total the value of total assets for all banks
as reported in the balance sheet. This will relate the leverage
measure to the reported total asset value for each bank. In this case for some
banks the leverage may be underestimated. The default value will be "asymmetric
residuals".}
}
\value{
A list, with the state variables at t = 0 and t = 1
        $e_0$ B x 1 vector of core tier 1 equity at t = 0
        $S_0$ B x I matrix of security exposures at t = 0
        $L_0$ B x J matrix of loan exposures at t = 0
        $e_1$ B x 1 vector of core tier 1 equity at t=1
        $S_1$ B x I matrix of security exposures at $t = 1$
        $L_1$ B x J matrix of loan exposures at $t=1$
        $theta$ B x 1 vector of fire sale proportions (a vector of zeros)
}
\description{
make_state_variables
}
\examples{
stress_data <- make_stress_data(eba_exposures_2016, eba_impairments_2016, 1, 2016)
          make_state_variables(stress_data, meth = "asymmetric residuals")
}
