% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_stress_data.R
\name{make_stress_data}
\alias{make_stress_data}
\title{make_stress_data}
\usage{
make_stress_data(exposure_data, impairment_data, horizon, base_year)
}
\arguments{
\item{exposure_data}{a dataframe with the exposure data. The exposure data
must contain the variables LEI_code (chr), Country_code (chr), Bank_name
(chr), Period (num), Country (chr), Exposure (chr), Loan_Amount (num),
Bond_Amount (num), Total_Amount (num), Unit (chr), Currency (chr)}

\item{impairment_data}{a dataframe with the impairment data. The impairment
data must contain the variables LEI_code (chr), Country_code (chr),
Bank_name (chr), Period (num), Scenario (num), Country (chr), Exposure
(chr), Impairment_rate (num)}

\item{horizon}{stress test horizon (1,2,3 years ahead)}

\item{base_year}{base year of the stress test (2015 or 2019)}
}
\value{
A dataframe with the following variables: LEI_code (chr), Country_code (chr),
  Bank_name (chr), Period (num), Country (chr), Exposure (chr), Loan_Amount (num), Bond_Amount (num),
  Total_Amount (num), Loan_Losses (num), Unit (chr), Currency (chr), Impairment_rate (num)
}
\description{
This function creates from exposure and impairment data a new dataframe which
makes all data available for conducting an EBA stress test, given two
parameters: \code{horizon}, the future horizon of losses assumed by the
stress test and \code{base_year}.
}
\examples{
make_stress_data(eba_exposures_2016, eba_impairments_2016, 1, 2015)
}
