% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_table_final_state.R
\name{make_table_final_state}
\alias{make_table_final_state}
\title{make_table_final_state}
\usage{
make_table_final_state(dat, fp, lb)
}
\arguments{
\item{dat}{a list of state matrices. The output of \code{make_state_variables}}

\item{fp}{a dataframe. The output of \code{fixed_point_computation_function}}

\item{lb}{a real number specifiying the leverage threshold. Same threshold as in
\code{fixed_point_computation_function}}
}
\value{
a B times 4 dataframe with the variables Bank_name, a_1_tau, e_1_tau,
and lambda_1_tau if the fixed point is unique a B times 7 dataframe with the variables
Bank_name, a_1_tau_lower, a_1_tau_upper, e_1_tau_lower, e_1_tau_lower and lambda_1_tau_lower,
lambda_1_tau_upper
}
\description{
This function takes the list of state values and the output of the
fixed point computation and generates a B times 4 dataframe with the
Bank_name, a_1_tau, e_1_tau, and lambda_1_tau
}
\examples{
stress_data <- make_stress_data(eba_exposures_2016, eba_impairments_2016, 1, 2015)
state_variables <- make_state_variables(stress_data)
fixed_points <- fixed_point_computation_function(
  mat = state_variables, lb = 33,
  data_idx = sovereign_bond_indices,
  data_adv = average_daily_volume_sovereign,
  base_year = 2015,
  constant = 1.5
)
make_table_final_state(state_variables, fixed_points, 33)
}
