% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_table_initial_state.R
\name{make_table_initial_state}
\alias{make_table_initial_state}
\title{make_table_initial_state}
\usage{
make_table_initial_state(dat)
}
\arguments{
\item{dat}{a list of state matrices. The output of \code{make_state_variables}}
}
\value{
a B times 4 dataframe with the variables Bank_name, a_0, e_0, and lambda_0
}
\description{
This function takes the list of state values and generates a B times 4 dataframe with the
Bank_name, a_0, e_0, and lambda_0
}
\examples{
stress_data <- make_stress_data(eba_exposures_2016, eba_impairments_2016, 1, 2015)
state_variables <- make_state_variables(stress_data)
make_table_initial_state(state_variables)
}
