% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_sovereign_bond_exposures.R
\name{prepare_sovereign_bond_exposures}
\alias{prepare_sovereign_bond_exposures}
\title{prepare_sovereign_bond_exposures}
\usage{
prepare_sovereign_bond_exposures(data)
}
\arguments{
\item{data}{a dataframe with the compiled EBA exposure data}
}
\value{
a dataframe with the bond exposures for DE, ES, IT, FR, UK, US and total
}
\description{
prepare_sovereign_bond_exposures
}
\examples{
prepare_sovereign_bond_exposures(eba_exposures_2016)
}
