% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/price_impact_function.R
\name{price_impact_function}
\alias{price_impact_function}
\title{price_impact_function}
\usage{
price_impact_function(
  del,
  mat,
  lb,
  data_idx,
  data_adv,
  base_year,
  constant,
  method = "squareroot"
)
}
\arguments{
\item{del}{I times 1 vector of discount factors with each component in
[0,1]}

\item{mat}{A list containing the matrices which are the output of
\code{make_matrices}.}

\item{lb}{A real number specifiying the leverage threshold lambd_bar.}

\item{data_idx}{A dataframe with price index data. It must contain the
variables Country (chr), Date, and Value (num)}

\item{data_adv}{A dataframe with average daily volumes. It must contain the
variables Country (chr), Year (num), Volume (num), Unit (chr) and Currency
(chr)}

\item{base_year}{The base year for the stress test exercise}

\item{constant}{The constant kappa in the square root formula for price
impact}

\item{method}{In our paper we use the square root law of price impact
because we can derive it from evidence and the market-microstructure
literature. But others, for example Schannning and Cont use different
impact functions. We leave space for allowing for different methods by
which impact is computed. For now we only implement the square root law.}
}
\value{
A list with An I times 1 vector of discount factors delta measuring
  the price impact of a quantity sold.
}
\description{
This function computes the price impact as a percentage loss in value of the
different bond exposures if a certain volume is sold to the market. It needs
to know the state variables, the data on bond volatilities and average daily
volumes as well as the base_year for which the stress test is taken as well
as the value of a constant of order unity. Finally it takes as argument a
method. Currently only one method is implemented, which we call
"square-root". This method implements equation (9) in the paper. For future
versions we keep space to include other impact functions which have been used
in the literature, in particular the functions of Cont and Schaanning 2017.
}
\examples{
stress_data <- make_stress_data(eba_exposures_2016, eba_impairments_2016, 1, 2015)
matrices <- make_state_variables(stress_data)
delta <- rep(0, dim(matrices$S_0)[2])
price_impact_function(delta,
matrices, 33,
sovereign_bond_indices,
average_daily_volume_sovereign,
2015,
1.5,
method = "squareroot")
}
