% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/price_impact_function_example.R
\name{price_impact_function_example}
\alias{price_impact_function_example}
\title{price_impact_function_example}
\usage{
price_impact_function_example(del, mat, lb)
}
\arguments{
\item{del}{I times 1 vector of discount factors with each component in $[0,1]$}

\item{mat}{A list containing the matrices which are the output of make_matrices}

\item{lb}{A real number specifiying the leverage threshold $bar{lambda}$}
}
\value{
A list containing the fixed point and the number of iterations needed to find it
}
\description{
price_impact_function_example
}
\examples{
price_impact_function_example(0, example_multiple_equilibria, 44)
}
