% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sov_bond_index_example.R
\docType{data}
\name{sov_bond_index_example}
\alias{sov_bond_index_example}
\title{A toy dataset for the example to illustrate the possibility of multiple fire sale equilibria.}
\format{
A 3 x 4 tibble with the following variables
\describe{
  \item{Country}{Usually an ISO code. Here the country is called Thomas}
  \item{Date}{A string. We use three ficticious 2020 dates}
  \item{Value}{Integers of ficticcious index value which will result in a volatility of approximately 1}

}
}
\source{
\url{Breuer, T. and Summer M., and Urosevic B., "Systemic Loss Evaluation", A3}
}
\usage{
sov_bond_index_example
}
\description{
This is a toy dataset for the example in Breuer, Summer, Urosevic, Appendix A3 given in a format
that it can be processed by the standard functions of the package which are also used for computing
the actual eba data
}
\keyword{datasets}
