% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sovereign_bond_indices.R
\docType{data}
\name{sovereign_bond_indices}
\alias{sovereign_bond_indices}
\title{Sovereign bond indices of DE, ES, FR, GB, IT, JP, US and a golbal world index.}
\format{
A data frame with 19733 rows and 3 variables:
\describe{
  \item{Country}{Country, ISO-code of the country.}
  \item{Date}{Date, date of the observation in the format mm/dd/yyyy}
  \item{Value}{Value, Index value}
}
}
\source{
\url{http://us.spindices.com}
}
\usage{
sovereign_bond_indices
}
\description{
A dataframe containing the daily values of sovereign composite debt indices for
Germany, Spain, France, Great Britain, Italy, Japan, United States and a
global index starting in 2010 up to the beginning of 2020.
}
\keyword{datasets}
