% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/average_daily_volume_sovereign.R
\docType{data}
\name{average_daily_volume_sovereign}
\alias{average_daily_volume_sovereign}
\title{Average daily volume in Million Euro of sovereign bonds of DE, ES, FR, IT, JP, UK, US.}
\format{
A data frame with 77 rows and 5 variables:
\describe{
  \item{Country}{Country, ISO-code of the country.}
  \item{Year}{Year, year of the observation}
  \item{Volume}{Volume, Average daily turnover volume in Million Euro}
  \item{Unit}{Unit, unit of Volume}
  \item{Currency}{Currency, currency of volume}
}
}
\source{
The volume data are handcollected from various sources on the internet
        DE \url{https://www.deutsche-finanzagentur.de/en/institutional-investors/secondary-market/}
        ES \url{https://www.tesoro.es/sites/default/files/estadisticas/15I.xlsx}
        FR \url{https://www.afme.eu/reports/data/details//Government-Bond-Data-Report-Q2-2019}
        IT \url{https://infostat.bancaditalia.it/}
        JP \url{https://asianbondsonline.adb.org/data-portal/}
        UK \url{https://www.dmo.gov.uk/data/gilt-market/turnover-data/}
        US \url{https://www.sifma.org/resources/research/us-treasury-trading-volume/}
}
\usage{
average_daily_volume_sovereign
}
\description{
A dataset containing the average daily volume of sovereign bonds for Germany, Spain, France, Italy, Japan,
United Kingdom, United States starting in 2009 up to the end of 2019.
}
\keyword{datasets}
