% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bank_behavior_function.R
\name{bank_behavior_function}
\alias{bank_behavior_function}
\title{bank_behavior_function}
\usage{
bank_behavior_function(del, mat, lb)
}
\arguments{
\item{del}{an I x 1 vector of discount factors with each entry in $[0,1]$}

\item{mat}{A list with all matrices. Initially the output of make_matrices}

\item{lb}{a real number specifying the leverage threshold a_1/e_1 above which banks start to sell bonds.}
}
\value{
A I x 1 vector of selling shares for all marketable asset classes.
}
\description{
bank_behavior_function
}
\examples{
stress_data <- make_stress_data(eba_exposures_2016, eba_impairments_2016, 1, 2015)
state_variables <- make_state_variables(stress_data)
bank_behavior_function(0, state_variables, 33)
}
