% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fixed_point_computation_function.R
\name{fixed_point_computation_function}
\alias{fixed_point_computation_function}
\title{fixed_point_computation_function}
\usage{
fixed_point_computation_function(
  del,
  mat,
  lb,
  data_idx,
  data_adv,
  base_year,
  constant,
  accuracy = 10^(-9)
)
}
\arguments{
\item{del}{A I time number of banks starting vector of discount factors}

\item{mat}{A list of the initial state variable}

\item{lb}{The critical leverage threshold called bar{lambda} in the paper.}

\item{data_idx}{The data-frame with the sovereign bond indices.}

\item{data_adv}{The data-frame with the average daily volume figures.}

\item{base_year}{The base year for the simulation}

\item{constant}{The value of the constant kappa in the impact fuction.}

\item{accuracy}{The accuracy of the fixed point approximation. Set by default to 10^9}
}
\value{
A list with components equ_delta (equilibrium discount factor) and iter (number of iterations)
}
\description{
fixed_point_computation_function
}
\examples{
fixed_point_computation_function(0,
  mat = example_multiple_equilibria, lb = 44,
  data_idx = sov_bond_index_example, data_adv = adv_example, base_year = 2020, constant = 0.9433962
)
}
