% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_stress_data.R
\name{make_stress_data}
\alias{make_stress_data}
\title{make_stress_data}
\usage{
make_stress_data(exposure_data, impairment_data, horizon, base_year)
}
\arguments{
\item{exposure_data}{a dataframe with the exposure data}

\item{impairment_data}{a dataframe with the impairment data}

\item{horizon}{stress test horizon (1,2,3 years ahead)}

\item{base_year}{base year of the stress test (for example 2015 for the 2016 stress test)}
}
\value{
A dataframe with the following variables:
        LEI_code, Country_code, Bank_name, Period, Country, Exposure,
        Loan_Amount, Bond_Amount, Total_Amount, Loan_Losses, Unit, Currency, Impairment_rate
}
\description{
This function creates from exposure and impairment data a new
dataframe which makes all data available for conducting an EBA stress test, given two parameters:
\code{horizon}, the future horizon of losses assumed by the stress test and \code{base_year}, the
base year when the stressed exposures are measured.
}
\examples{
make_stress_data(eba_exposures_2016, eba_impairments_2016, 1, 2015)
}
