% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/price_impact_function.R
\name{price_impact_function}
\alias{price_impact_function}
\title{price_impact_function}
\usage{
price_impact_function(
  del,
  mat,
  lb,
  data_idx,
  data_adv,
  base_year,
  constant,
  method = "squareroot"
)
}
\arguments{
\item{del}{I times 1 vector of discount factors with each component in $[0,1]$}

\item{mat}{A list containing the matrices which are the output of make_matrices}

\item{lb}{A real number specifiying the leverage threshold $bar{lambda}$}

\item{data_idx}{A dataframe with price index data}

\item{data_adv}{A dataframe with average daily volumes}

\item{base_year}{The base year for the stress test exercise}

\item{constant}{The constant $kappa$ in the square root formula for price impact}

\item{method}{In our paper we use the suqare root law of price impact because we can derive it
from evidence and the market-microstructure literature. But others, for example Schannning and Cont
use different impact functions. We leave space for allowing for different methods by which
impact is computed. For now we only implement the square root law.}
}
\value{
A list with An I times 1 vector of discount factors $delta$ measuring the price impact of
        a quantity sold and a T times 1 vector of maximum impact (the impact when all bonds held by the banks
        together were sold off.)
}
\description{
price_impact_function
}
\examples{
price_impact_function(0, example_multiple_equilibria, 44, sov_bond_index_example,
                                adv_example, 2020, 0.9433962)
}
